refactor: Fix quadruple line break removal in chapter detail
The code changes in this commit fix the regular expression used to remove quadruple line breaks in the chapter detail. The previous regular expression was not accounting for spaces between the line breaks, resulting in incorrect removal. This fix ensures that quadruple line breaks are correctly replaced with double line breaks.